PM-443 Decoupling tests by user feature in selectors
